% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_wrapper.R
\name{install_python_environment}
\alias{install_python_environment}
\title{Install a python environment - run once after initial installation}
\usage{
install_python_environment()
}
\value{
No return but a python environment will be installed via miniconda at r-reticulate
}
\description{
Install a python environment - run once after initial installation
}
\examples{
install_python_environment()
}
