% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_wrapper.R
\name{make_cube}
\alias{make_cube}
\title{Creates a spacetime cube object and writes a .nc4 file to disk}
\usage{
make_cube(...)
}
\arguments{
\item{data}{A spacetime file object}

\item{fileName}{The intended name of the file written to disk (character string)}

\item{organizeFiles}{How files are intended to be organized in the cube (character string)}

\item{organizeBands}{How bands are intended to be organized in the cube (character string)}

\item{varNames}{A list of variable names (list)}

\item{timeObj}{A spacetime time object}

\item{inMemory}{Should the operation be virtualized (slower but easier on memory) or done in memory (faster but higher memory usage) (boolean or character)}
}
\value{
A spacetime cube object and a nc4 file written to disk
}
\description{
Creates a spacetime cube object and writes a .nc4 file to disk
}
\examples{
make_cube(data = fileObj, fileName = file.nc4, organizeFiles="filestotime", organizeBands="bandstotime")
}
