% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_wrapper.R
\name{raster_align}
\alias{raster_align}
\title{Aligns rasters in a file object to have the same pixel size, alignment, and spatial reference system}
\usage{
raster_align(...)
}
\arguments{
\item{data}{A spacetime file object}

\item{noneVal}{The intended no data value (numeric) Defaults to the first in the stack}

\item{resolution}{The resolution of the intended output (character string or numeric). "min" (selects the largest pixel size in the stack - default), "max" (smallest pixel size in stack), numeric value (the intended pixel size in the units of the intended SRS). "}

\item{SRS}{The spatial reference system/EPSG code (numeric) Defaults to the first in the stack}

\item{algorithm}{The aggregation algorithm method (character string)}
}
\value{
A spacetime file object with aligned files
}
\description{
Aligns rasters in a file object to have the same pixel size, alignment, and spatial reference system
}
\examples{
raster_align(data = fileObj, SRS = 4326, noneVal = -9999)
}
