% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_wrapper.R
\name{raster_trim}
\alias{raster_trim}
\title{Trims rasters in a file object in the manner specified by the "method" argument}
\usage{
raster_trim(...)
}
\arguments{
\item{data}{A spacetime file object}

\item{ul}{A list of lat and lon values for the upper left corner of a bounding box (list)}

\item{lr}{A list of lat and lon values for the lower right corner of a bounding box (list)}

\item{method}{The method used to trim raster stack. Options are "intersection", "union", "corners" or "shape" (character string)}

\item{shapeFile}{The name of or path to a shape file (character string)}
}
\value{
A spacetime file object with extents trimmed
}
\description{
Trims rasters in a file object in the manner specified by the "method" argument
}
\examples{
raster_trim(data = fileObj, shapeFile = "file.shp", method = "shape")
}
