% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_wrapper.R
\name{read_data}
\alias{read_data}
\title{Read in raster data as a spacetime file object}
\usage{
read_data(...)
}
\arguments{
\item{data}{A list of raster files}
}
\value{
A spacetime file object
}
\description{
Read in raster data as a spacetime file object
}
\examples{
read_data(data = data_list)
}
